community-service/requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
psycopg2
pika
pillow
django-cors-headers
python-dotenv
requests
channels[daphne]
channels_redis